Adds tool based warning retract. amount and speed.
Two new private variables "_retraction_amount_max_warning" and "_retraction_speed_max_warning" were introduced. These are used by the different variant files to set the maximum retraction speed and amount.

The main goal is to prevent user form settings these values to high and thus causing the system to clog. That is also why only a maximum for the warning is set.
-